Extract functions to separate file
